Add tests to subroutines
This exposed a number of changes that needed to be made: the big one was
that I needed to support explicit seeding and RNG generator creation to
be able to test the pseudorandomness of my generators.

There are still a few issues: the tests are WAY too low (I think due to
environment creation having too much overhead?) and make_envs is missing
tests at the moment.