Add greedy policies and q-function update
This commit is larger than I expected, since it exposed the fact that
I shouldn't use the RNG from OpenAI Gym in the random policies (so that
they can be inline with the RNGs in the greedy algorithms).

The good news is that the policy tests are much faster now.  The bad
news is that the API is a mess now, and there's too many hardcoded
values throughout the tests and code that needs to be minimized.

I've held off on commiting the example for running the greedy algorithm
until I've sanitized the code.