(WIP) Move control to classes
While thinking about how to do hyperparameter optimization, I've come to
the conclusion that I need to re-implement the updates as classes to
standardize the API for hyperparameter optimization and suitable output
to pandas.

I'm not close to done, but I've gone way too long without a commit.

Current issues are:

* Base class needs to be created.
* Docstrings haven't been implemented for new classes.
* Ensure my naming scheme/terminology matches the literature.
* Changes haven't been merged into the trials (I will likely be
  completely redoing the trials once everything is said and done.)