Add notebook for Deep-Q with Atari
This is my attempt to recreate the 2015 DeepMind paper.  I've been
working on it for a bit, but have decided to incorporate it into this
repo to get it under version control.

The current notebook is using Breakout, since it's a pretty simple game
that I'm familiar with.  The networks are written in PyTorch, since I
prefer its coding aesthetic more than TensorFlow.

The training loop is currently unruly, and the training is suffering
from catastrophic forgetting.