Move (most) PyTorch functionality into model.py
The two major exceptions are some torch.save() calls in the QLearning
save(), which don't really affect the overall structure so I'll leave
them be for now, and the specification of "device" throughout
train().  This isn't *technically* PyTorch functionality, but it's
getting close, and I don't feel that this belongs in the main
Q-learning loop.